requirements.txt
pandas
numpy
psycopg2
nbformat
python-dotenv
typing_extensions